Add openssl as a recommended dependency
`hab` uses the OpenSSL library in order to make secure HTTPS
connections to the Builder service. In order to do this, it must have
access to the necessary certificates in order to establish trust. The
easiest way to do this on MacOS is to use Homebrew to also install
`openssl`, which generates a `/usr/local/etc/openssl/cert.pem` file
based on the contents of the SystemRoots keychain.

If this file is not present, users will see the following error
message (wrapped for presentation purposes):

```
the handshake failed: The OpenSSL library reported an error:
error:14090086:SSL routines:ssl3_get_server_certificate:certificate
verify failed:s3_clnt.c:1264:: unable to get local issuer certificate
```

In this commit, `openssl` is added as a "recommended" Homebrew
dependency. This means that it will be installed by default when a
user runs `brew install hab`. Should the user not want to install
`openssl` from Homebrew, they can install `hab` like this:

```sh
brew install hab --without-openssl
```

They will then be responsible for telling `hab` where it can find
certificates, using the OpenSSL standard environment variables
`SSL_CERT_FILE` and/or `SSL_CERT_DIR`.

Signed-off-by: Christopher Maier <cmaier@chef.io>